refactor: Notice 생성하는 파일들 이름 변경
