+++ src/main/java/com/real/backend/domain/notice/controller/NoticeController.java
@@ -1,5 +1,7 @@
 package com.real.backend.domain.notice.controller;
 
+import java.util.List;
+
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,14 +11,16 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.real.backend.domain.notice.dto.NoticeCreateRequestDTO;
 import com.real.backend.domain.notice.dto.NoticeInfoResponseDTO;
 import com.real.backend.domain.notice.dto.NoticeListResponseDTO;
 import com.real.backend.domain.notice.service.NoticeService;
-import com.real.backend.domain.notice.tmp.NoticeCreateRequestTmpDTO;
+import com.real.backend.domain.notice.dto.NoticePasteRequestTmpDTO;
 import com.real.backend.response.DataResponse;
 import com.real.backend.response.StatusResponse;
 import com.real.backend.security.CurrentSession;
@@ -68,7 +72,7 @@ public class NoticeController {
         return DataResponse.of(noticeInfoResponseDTO);
     }
 
-    // @PreAuthorize("!hasAnyAuthority('OUTSIDER', 'TRAINEE')")
+    @PreAuthorize("!hasAnyAuthority('OUTSIDER', 'TRAINEE')")
     @DeleteMapping("/v1/notices/{noticeId}")
     public StatusResponse deleteNotice(
         @PathVariable Long noticeId,
@@ -78,14 +82,24 @@ public class NoticeController {
         return StatusResponse.of(204, "공지가 성공적으로 삭제되었습니다.");
     }
 
-    // @PreAuthorize("!hasAnyAuthority('OUTSIDER', 'TRAINEE')")
+    @PreAuthorize("!hasAnyAuthority('OUTSIDER', 'TRAINEE')")
     @PutMapping("/v1/notices/{noticeId}")
     public StatusResponse editNotice(
         @PathVariable Long noticeId,
-        @RequestBody NoticeCreateRequestTmpDTO noticeCreateRequestTmpDTO,
+        @RequestBody NoticePasteRequestTmpDTO noticePasteRequestTmpDTO,
         @CurrentSession Session session
     ) {
-        noticeService.editNotice(noticeId, noticeCreateRequestTmpDTO);
+        noticeService.editNotice(noticeId, noticePasteRequestTmpDTO);
         return StatusResponse.of(200, "공지가 성공적으로 수정되었습니다.");
+    }
+
+    @PostMapping("/v1/notices/tmp")
+    public StatusResponse pasteNoticeTmp(
+        @RequestPart("notice") NoticePasteRequestTmpDTO noticePasteRequestDTO,
+        @RequestPart(value = "images", required = false) List<MultipartFile> images,
+        @RequestPart(value = "files", required = false) List<MultipartFile> files
+    ) throws JsonProcessingException {
+        noticeService.pasteNoticeTmp(noticePasteRequestDTO, images, files);
+        return StatusResponse.of(201, "공지가 성공적으로 생성되었습니다.");
     }
 }
