+++ src/main/java/com/real/backend/domain/notice/domain/Notice.java
@@ -3,7 +3,7 @@ package com.real.backend.domain.notice.domain;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.real.backend.domain.notice.tmp.NoticeCreateRequestTmpDTO;
+import com.real.backend.domain.notice.dto.NoticePasteRequestTmpDTO;
 import com.real.backend.domain.user.domain.User;
 import com.real.backend.domain.user.domain.UserNoticeRead;
 import com.real.backend.post.Post;
@@ -43,14 +43,14 @@ public class Notice extends Post {
     @OneToMany(mappedBy = "notice", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<UserNoticeRead> userNoticeReads;
 
-    public void updateNotice(NoticeCreateRequestTmpDTO noticeCreateRequestTmpDTO, User user) {
-        this.originalUrl = noticeCreateRequestTmpDTO.originalUrl();
-        this.platform = noticeCreateRequestTmpDTO.platform();
+    public void updateNotice(NoticePasteRequestTmpDTO noticePasteRequestTmpDTO, User user) {
+        this.originalUrl = noticePasteRequestTmpDTO.originalUrl();
+        this.platform = noticePasteRequestTmpDTO.platform();
         this.user = user;
         this.updatePost(
-            noticeCreateRequestTmpDTO.title(),
-            noticeCreateRequestTmpDTO.content(),
-            noticeCreateRequestTmpDTO.tag()
+            noticePasteRequestTmpDTO.title(),
+            noticePasteRequestTmpDTO.content(),
+            noticePasteRequestTmpDTO.tag()
             );
     }
 }
