+++ src/main/java/com/real/backend/domain/notice/service/NoticeService.java
@@ -3,16 +3,21 @@ package com.real.backend.domain.notice.service;
 import static com.real.backend.util.CursorUtils.*;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.real.backend.domain.notice.domain.Notice;
 import com.real.backend.domain.notice.component.NoticeFinder;
-import com.real.backend.domain.notice.tmp.NoticeCreateRequestTmpDTO;
+import com.real.backend.domain.notice.domain.NoticeFile;
+import com.real.backend.domain.notice.dto.NoticePasteRequestTmpDTO;
+import com.real.backend.domain.notice.repository.NoticeFileRepository;
 import com.real.backend.domain.user.domain.UserNoticeRead;
 import com.real.backend.domain.notice.dto.NoticeCreateRequestDTO;
 import com.real.backend.domain.notice.dto.NoticeFileGroups;
@@ -28,6 +33,7 @@ import com.real.backend.infra.ai.dto.NoticeSummaryRequestDTO;
 import com.real.backend.infra.ai.dto.NoticeSummaryResponseDTO;
 import com.real.backend.infra.ai.service.NoticeAiService;
 import com.real.backend.util.CursorUtils;
+import com.real.backend.util.S3Utils;
 import com.real.backend.util.dto.SliceDTO;
 
 import lombok.RequiredArgsConstructor;
@@ -35,14 +41,16 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class NoticeService {
-    private final NoticeRepository noticeRepository;
     private final NoticeLikeService noticeLikeService;
     private final NoticeFileService noticeFileService;
     private final NoticeAiService noticeAiService;
+    private final UserRepository userRepository;
+    private final NoticeRepository noticeRepository;
     private final UserNoticeReadRepository userNoticeReadRepository;
+    private final NoticeFileRepository noticeFileRepository;
     private final NoticeFinder noticeFinder;
     private final UserFinder userFinder;
-    private final UserRepository userRepository;
+    private final S3Utils s3Utils;
 
     @Transactional(readOnly = true)
     public SliceDTO<NoticeListResponseDTO> getNoticeListByCursor(Long cursorId, int limit, String cursorStandard, Long userId) {
@@ -134,11 +142,89 @@ public class NoticeService {
     }
 
     @Transactional
-    public void editNotice(Long noticeId, NoticeCreateRequestTmpDTO noticeCreateRequestTmpDTO) {
+    public void editNotice(Long noticeId, NoticePasteRequestTmpDTO noticePasteRequestTmpDTO) {
         Notice notice = noticeFinder.getNotice(noticeId);
         User user = userRepository.findById(notice.getUser().getId()).orElseThrow(() -> new NotFoundException("User not found"));
 
-        notice.updateNotice(noticeCreateRequestTmpDTO, user);
+        notice.updateNotice(noticePasteRequestTmpDTO, user);
         noticeRepository.save(notice);
+    }
+
+    @Transactional
+    public void pasteNoticeTmp(NoticePasteRequestTmpDTO noticeCreateRequestDTO, List<MultipartFile> images,
+        List<MultipartFile> files) throws JsonProcessingException {
+
+        String userName = noticeCreateRequestDTO.userName();
+        User user = userRepository.findByNickname(userName);
+
+        // ai에 summary 요청 로직
+        // NoticeSummaryRequestDTO noticeSummaryRequestDTO = new NoticeSummaryRequestDTO(noticeCreateRequestDTO.content(),
+        //     noticeCreateRequestDTO.title());
+        // NoticeSummaryResponseDTO noticeSummaryResponseDTO = null;
+        // for (int i = 0; i < 3; i++) {
+        //     noticeSummaryResponseDTO = noticeAiService.makeSummary(noticeSummaryRequestDTO);
+        //     if (noticeSummaryResponseDTO.isCompleted())
+        //         break;
+        // }
+        // if (!noticeSummaryResponseDTO.isCompleted()){
+        //     throw new ServerException("ai가 응답을 주지 못했습니다.");
+        // }
+
+        LocalDateTime createTime = LocalDateTime.parse(noticeCreateRequestDTO.createdAt());
+
+        Notice notice = noticeRepository.save(Notice.builder()
+            .user(user)
+            .title(noticeCreateRequestDTO.title())
+            .content(noticeCreateRequestDTO.content())
+            // .summary(noticeSummaryResponseDTO.summary())
+            .platform(noticeCreateRequestDTO.platform())
+            .tag(noticeCreateRequestDTO.tag())
+            .originalUrl(noticeCreateRequestDTO.originalUrl())
+            .totalViewCount(0L)
+            .commentCount(0L)
+            .likeCount(0L)
+            .createdAt(createTime)
+            .build());
+
+        int i = 0;
+        if (images == null)
+            images = new ArrayList<>();
+        if (files == null)
+            files = new ArrayList<>();
+        for (MultipartFile file : files) {
+            String url = s3Utils.upload(file, "static/notice/files");
+            String name = file.getOriginalFilename();
+            String type = "";
+            if (name != null && name.contains(".")) {
+                type = name.substring(name.lastIndexOf(".") + 1).toLowerCase();
+            }
+            i++;
+            Integer fileSeqNo = i;
+            noticeFileRepository.save(NoticeFile.builder()
+                .notice(notice)
+                .fileSeqNo(fileSeqNo)
+                .fileUrl(url)
+                .type(type)
+                .build());
+
+        }
+        i = 0;
+        for (MultipartFile file : images) {
+            String url = s3Utils.upload(file, "static/notice/images");
+            String name = file.getOriginalFilename();
+            String type = "";
+            if (name != null && name.contains(".")) {
+                type = name.substring(name.lastIndexOf(".") + 1).toLowerCase();
+            }
+            Integer fileSeqNo = i;
+            i++;
+            noticeFileRepository.save(NoticeFile.builder()
+                .notice(notice)
+                .fileSeqNo(fileSeqNo)
+                .fileUrl(url)
+                .type(type)
+                .build());
+
+        }
     }
 }
